+++ src/components/features/Header.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 export const HEADER_HEIGHT = "64px";
 const Header = () => {
   return <div>Header</div>;
